Code Maintenance - Error Prone has been added to the build   * Fixed all the issues found - Turned NullAway on then turned it off since it crashes due to Java 14 issues - Updated Concurrent auto code generation - Removed auto generated MT files from repo - Removed auto generated F32 files from repo - PixelMath  * Added Lambda functions

+++ main/boofcv-feature/src/main/java/boofcv/alg/descriptor/DescriptorDistance.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2019, Peter Abeles. All Rights Reserved.
+ * Copyright (c) 2011-2020, Peter Abeles. All Rights Reserved.
  *
  * This file is part of BoofCV (http://boofcv.org).
  *
@@ -74,7 +74,7 @@ public class DescriptorDistance {
 		final int N = a.value.length;
 		float total = 0;
 		for( int i = 0; i < N; i++ ) {
-			double d = a.value[i]-b.value[i];
+			float d = a.value[i]-b.value[i];
 			total += d*d;
 		}
